Refactor functions and lift tour step
Refactored some functionalities since they could be put into a util
components. Refactor Tour steps to be present in parent component more.
